Fix compiler warning C4244: 'argument' : conversion from 'DWORD' to 'BYTE', possible loss of data
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@7054 3231ed4b-2112-0410-84e0-8c3944a40730

+++ shareaza/CoolInterface.cpp
@@ -393,7 +393,7 @@ void CCoolInterface::CreateFonts(LPCTSTR pszFace, int nSize)
 	if ( m_fntBoldItalic.m_hObject ) m_fntBoldItalic.DeleteObject();
 	if ( m_fntNavBar.m_hObject ) m_fntNavBar.DeleteObject();
 
-	DWORD nQuality = theApp.m_dwWindowsVersion >= 6 ? DEFAULT_QUALITY : ANTIALIASED_QUALITY;
+	BYTE nQuality = theApp.m_dwWindowsVersion >= 6 ? DEFAULT_QUALITY : ANTIALIASED_QUALITY;
 
 	m_fntNormal.CreateFontW( -nSize, 0, 0, 0, FW_NORMAL, FALSE, FALSE, FALSE,
 		DEFAULT_CHARSET, OUT_DEFAULT_PRECIS, CLIP_DEFAULT_PRECIS, nQuality,
